requirements.txt
openai>=1.12.0
sqlalchemy>=2.0.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0  # For environment variables
tenacity>=8.0.0  # For retries
Pillow
